Move compare_quant and compare_sym_asym to common_comparers
